requirements.txt
# ============================================
# AI Financial Coach - Python Dependencies
# ============================================

# Core Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart

# Database
motor>=3.6.0
pymongo>=4.9,<4.10

# Data Validation & Schemas
pydantic>=2.9.0
pydantic-settings>=2.6.0

# LangChain & LangGraph (Multi-Agent Orchestration)
langgraph>=0.2.40
langchain>=0.3.0
langscikit-learn>=1.3.0

# === FastAPI Backend (Phase 6) ===
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads
langchain-openai>=0.2.0
langchain-community>=0.3.0  # Required by PaddleOCR

# LLM API
openai>=1.50.0

# OCR - PaddleOCR
paddlepaddle>=2.6.0
paddleocr>=2.7.0
protobuf

# PDF Processing
pdf2image
pdfplumber
PyPDF2
pillow

# Data Processing (numpy<2 for langchain compatibility)
pandas
numpy<2,>=1.26.0

# Environment & Configuration
python-dotenv

# HTTP Requests
httpx
aiohttp

# Logging & Monitoring
loguru

# Testing
pytest
pytest-asyncio

# CORS & Security
python-jose[cryptography]

# Utilities
python-dateutil
